+++ src/main/kotlin/uk/gov/justice/digital/hmpps/visitscheduler/service/ApplicationService.kt
@@ -502,6 +502,6 @@ class ApplicationService(
   @Transactional
   fun updateApplicationsPrisonerIdPostMerge(oldPrisonerId: String, newPrisonerId: String) {
     // update all applications from oldPrisonerId to newPrisonerId post merge
-    applicationRepository.updateApplication(oldPrisonerId = oldPrisonerId, newPrisonerId = newPrisonerId)
+    applicationRepository.updatePrisonerId(oldPrisonerId = oldPrisonerId, newPrisonerId = newPrisonerId)
   }
 }
